+++ packages/adminui-vue/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - `UTableEntity` - fixed hiding the filter when the table has one column and fix entity-equal filter
 
 ## [5.22.4] - 2022-02-16
 ### Added